Move Checkbox to common-ui
