+++ src/common-ui/components/index.js
@@ -12,3 +12,5 @@ export { default as IndexDropdownRow } from './IndexDropdownRow'
 export { default as IndexDropdownNewRow } from './IndexDropdownNewRow'
 export { default as ProgressBar } from './ProgressBar'
 export { default as ToggleSwitch } from './ToggleSwitch'
+export { default as Checkbox } from './Checkbox'
+export { CheckboxToggle } from './Checkbox'
